First commit
Tested and seems to work perfectly.  Room for improvement would be in the testing for start and end time to make so as if the power is interrupted, then it would test if it needs to be on or off and resume as normal.  Could also maybe add lib for time zones and used for automatically setting DST or not based on named location.